// frontend/src/types/generated/ReleaseEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AutoName } from "./AutoName";
import type { FarmStage } from "./FarmStage";
import type { OpStage } from "./OpStage";
import type { Resolution } from "./Resolution";

export type ReleaseEvent = { cnId: string; nameCn: string; nameEn: string | null; activityType: string; hasStage: boolean; cnStart: number; cnEnd: number; opStages: Array<OpStage>; farmStages: Array<FarmStage>; nameEnAuto: AutoName | null; imagePath: string | null; resolution: Resolution };